renamed reducer
