+++ src/Components/PackageOverview/PackageOverview.jsx
@@ -5,8 +5,8 @@ import { useHistory } from "react-router-dom";
 
 import Button from "../Button/Button.jsx";
 
-import "../../redux/Actions/learn.js";
-import { setLearnedPackage } from "../../redux/Actions/learn.js";
+import "../../redux/Actions/query.js";
+import { setLearnedPackage } from "../../redux/Actions/query.js";
 
 import "./PackageOverview.scss";
