+++ src/redux/Reducers/index.js
@@ -1,8 +1,8 @@
 import { combineReducers } from "redux";
 
 import formReducer from "./form.js";
-import learnReducer from "./learn.js";
 import loginReducer from "./login.js";
+import queryReducer from "./query.js";
 import settingReducer from "./setting.js";
 import tableReducer from "./table.js";
 
@@ -11,7 +11,7 @@ const allReducers = combineReducers({
   setting: settingReducer,
   form: formReducer,
   table: tableReducer,
-  learn: learnReducer,
+  query: queryReducer,
 });
 
 export default allReducers;
